requirements.txt
asgiref==3.2.7
astroid==2.4.1
dj-database-url==0.5.0
Django==3.0.7
django-extensions==2.2.9
django-filter==2.2.0
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
pylint==2.5.2
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
